+++ embark/workers/urls.py
@@ -14,8 +14,10 @@ urlpatterns = [
     path('worker/configure/<int:configuration_id>/', views.configure_worker, name='embark-worker-configure'),
     path('worker/connect/<int:configuration_id>/<int:worker_id>/', views.connect_worker, name='embark-worker-connect'),
     path('worker/registered/<int:configuration_id>/', views.registered_workers, name='embark-worker-registered'),     # TODO: convert to API endpoints after development complete
-    path('worker/soft_reset/<int:worker_id>/', views.worker_soft_reset, name='embark-worker-soft-reset'),     # TODO: add HTML context for this
-    path('worker/soft_reset/<int:worker_id>/<int:configuration_id>/', views.worker_soft_reset, name='embark-worker-soft-reset'),     # TODO: add HTML context for this
+    path('worker/soft_reset/<int:worker_id>/', views.worker_soft_reset, name='embark-worker-soft-reset'),
+    path('worker/soft_reset/<int:worker_id>/<int:configuration_id>/', views.worker_soft_reset, name='embark-worker-soft-reset'),
+    path('worker/hard_reset/<int:worker_id>/', views.worker_hard_reset, name='embark-worker-hard-reset'),
+    path('worker/hard_reset/<int:worker_id>/<int:configuration_id>/', views.worker_hard_reset, name='embark-worker-hard-reset'),
     path('worker/update/<int:worker_id>/', views.update_worker_dependency, name='embark-worker-update'),
     path('worker/update/configuration/<int:configuration_id>/', views.update_configuration_dependency, name='embark-configuration-update'),
 ]